lua: lua_field_inspect_table without pushcfunction
Currently on encoding table we push cfunction (lua_field_try_serialize)
to lua stack with additional lightuserdata and table value and after
pcall that function to avoid a raise of error.

In this case LuaJIT creates new object which will not live long time,
so it increase amount of dead object and also increase time and
frequency of garbage collection inside LuaJIT.
Also this pcall is necessary only in case when metafield __serialize
of serilizable object has LUA_TFUNCTION type.

So instead pushcfunction with pcall we can directly call the function
trying to serialize an object.